rag_service/requirements.txt
fastapi
uvicorn
pydantic
supabase
google-generativeai
groq
python-dotenv
PyPDF2
langchain-text-splitters
python-multipart